refactor: utils 전역 파일로 이동 & 필드네이밍 수정
